+++ source/Engine/Rendering/LineStyle.ts
@@ -2,5 +2,6 @@ export interface LineStyle {
     color?: string,
     lineCap?: CanvasLineCap,
     lineJoin?: CanvasLineJoin,
-    lineWidth?: number
+    lineWidth?: number,
+    dashed?: boolean
 }