+++ source/Maps/Contents/Configuration/ContentPointConfiguration.ts
@@ -4,5 +4,6 @@ import { ContentPointConstraint } from "./ContentPointConstraint";
 
 export interface ContentPointConfiguration {
     type: ContentPointType;
+    connections: number[];
     constraints: ContentPointConstraint[];
 }