+++ source/Maps/Contents/ContentPoint.ts
@@ -5,5 +5,6 @@ import { ContentPointType } from "./ContentPointType";
 export interface ContentPoint {
     type: ContentPointType,
     point: Point,
+    connections: number[],
     constraints?: ContentPointConstraint[]
 }