+++ CHANGELOG.md
@@ -12,7 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Tools:
   - **River**: The tool tries to avoid creating loops when generating a river
   - **Road**: Roads are created cell-by-cell instead of on a straight line
-  - **Select**: Neighboring cells are shown while editing a cell
+  - **Select**:
+      - Neighboring cells are shown while editing a cell
+      - A dashed line connects a helper point with its relevant primary point
 
 ### Fixed